chore: refactored sqlglot toolbox configuration and folder structure
