+++ .gitattributes
@@ -1,3 +1,3 @@
 poetry.lock linguist-generated=true
-exasol/exasol.sqlglot_toolbox/version.py linguist-generated=true
+sqlglot_toolbox/version.py linguist-generated=true
 doc/changes/changelog.md linguist-generated=true
