+++ __init__.py
@@ -1,3 +1,3 @@
-from exasol.sqlglot_toolbox.dialects.exasol import ExasolToolBox # noqa: F401
+from sqlglot_toolbox.dialects.exasol import ExasolToolBox # noqa: F401
