+++ noxfile.py
@@ -4,4 +4,4 @@ import nox
 from exasol.toolbox.nox.tasks import *
 
 # default actions to be run if nothing is explicitly specified with the -s option
-nox.options.sessions = ["project:fix"]
+nox.options.sessions = ["test:unit", "test:coverage"]
