+++ sqlglot_toolbox/dialects/exasol.py
