+++ test/unit/unit_smoke_test.py
@@ -1,5 +1,2 @@
-import exasol.sqlglot_toolbox
-
-
 def test_unit_smoke_test():
     assert True
